+++ src/Block/ResultWriter/FilenameProvider/LiteralFilenameProvider.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Sobak\Scrawler\Block\ResultWriter\FilenameProvider;
+
+class LiteralFilenameProvider extends AbstractFilenameProvider
+{
+    public function generateFilename(object $entity): string
+    {
+        return $this->configuration['filename'];
+    }
+}